servicenow_agent/requirements.txt
google-adk>=1.8.0
requests>=2.31.0
vertexai>=1.38.0
python-dotenv>=1.0.0
